@@ -15,8 +15,9 @@ require 'solidus_dev_support/rspec/feature_helper'
 # in spec/support/ and its subdirectories.
 Dir[File.join(File.dirname(__FILE__), 'support/**/*.rb')].each { |f| require f }
 
-# Requires factories defined in lib/solidus_sale_prices/factories.rb
-require 'solidus_sale_prices/factories'
+# Will load Solidus core factory first and then the ones
+# defined in `lib/solidus_sale_prices/testing_support/factories`.
+SolidusDevSupport::TestingSupport::Factories.load_for(SolidusSalePrices::Engine)
 
 RSpec.configure do |config|
   config.infer_spec_type_from_file_location!
